migrationエラーを修正

+++ db/schema.rb
@@ -18,7 +18,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_03_30_055354) do
     t.integer "quantity", default: 1, null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.integer "cart_id", null: false
+    t.bigint "cart_id", null: false
     t.bigint "item_id", null: false
     t.index ["cart_id"], name: "index_cart_items_on_cart_id"
     t.index ["item_id"], name: "index_cart_items_on_item_id"
